+++ spec/cbench/cbench_switch_spec.rb
@@ -1,5 +1,0 @@
-require 'trema'
-require 'cbench_switch'
-
-describe CbenchSwitch do
-end