+++ CHANGELOG.md
@@ -3,13 +3,17 @@
 ## develop (unreleased)
 
 
-## 0.1.1 (2/6/2015)
+## 0.2.0 (2/14/2015)
+### Changes
+* Add `CbenchSwitch#start` handler.
+* Use `-p`, `-l` and `-s` options of `trema run` in cucumber.
 
+
+## 0.1.1 (2/6/2015)
 ### Bugs fixed
 * Automatically respond to Echo request messages from switches.
 
 
 ## 0.1.0 (2/5/2015)
-
 ### Misc
 * The initial release version of hello_trema that runs on [pure-Ruby Trema](https://github.com/trema/trema_ruby).